wordcount example replace for clarity
